backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Bcrypt
python-dotenv